refactor(desktop): 抽离 useGlobalShortcuts，App 卸下全局快捷键逻辑
把窗口级 keydown（F5 自动评审 / DevTools / 查看已关闭 / Ctrl-Cmd+B·J 布局开关）
及配套 selectedId/canEngage ref 收进 hooks/useGlobalShortcuts，App 仅保留一行装配。
零行为变更。
